frontend/requirements.txt
fastapi==0.115.1
streamlit==1.47.0
python-multipart==0.0.7
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.3
pydantic==2.10.6
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2